+++ mods/BrewLAN/modules/BrewUI/strategicicons/hook/lua/system/Blueprints.lua
@@ -29,8 +29,6 @@ end
 local function unitHasWallCat(bp)
     local t = {
         WALL = true,
-        MEDIUMWALL = true,
-        HEAVYWALL = true,
         SHIELDWALL = true,
     }
     if bp.Categories then